refactor: implement hooks as helpers
