+++ src/components/FieldAdapter.js
@@ -1,29 +1,24 @@
-import React, { useCallback } from 'react'
+import React from 'react'
 import propTypes from '@dhis2/prop-types'
 import { fieldRenderProps, fieldInputProps, fieldMetaProps } from './Field'
 
 const PRIMITIVE_TYPES = new Set(['string', 'number', 'boolean'])
 
-const useChangeHandler = onChange =>
-    useCallback(
-        payload => {
-            const value =
-                // ui-core event signature
-                (payload && payload.value) ||
-                // synthetic event
-                (payload && payload.target && payload.target.value) ||
-                // directly usable value
-                (PRIMITIVE_TYPES.has(typeof payload) && payload) ||
-                // ¯\_(ツ)_/¯
-                ''
+const createChangeHandler = onChange => payload => {
+    const value =
+        // ui-core event signature
+        (payload && payload.value) ||
+        // synthetic event
+        (payload && payload.target && payload.target.value) ||
+        // directly usable value
+        (PRIMITIVE_TYPES.has(typeof payload) && payload) ||
+        // ¯\_(ツ)_/¯
+        ''
 
-            onChange(value)
-        },
-        [onChange]
-    )
+    onChange(value)
+}
 
 const FieldAdapter = ({ input, meta, component: Component, ...ownProps }) => {
-    const handleChange = useChangeHandler(input.onChange)
     const error = ownProps.error || (meta.touched && meta.invalid)
 
     return (
@@ -43,7 +38,7 @@ const FieldAdapter = ({ input, meta, component: Component, ...ownProps }) => {
             validationText={
                 ownProps.validationText || (error && meta.error) || ''
             }
-            onChange={handleChange}
+            onChange={createChangeHandler(input.onChange)}
         />
     )
 }
