+++ src/components/MultiSelect.js
@@ -1,34 +1,27 @@
-import React, { useCallback, useMemo } from 'react'
+import React from 'react'
 import propTypes from '@dhis2/prop-types'
 import { MultiSelectField, MultiSelectOption } from '@dhis2/ui-core'
 
 import { FieldAdapter } from './FieldAdapter.js'
 
-const useChangeHandler = onChange =>
-    useCallback(
-        ({ selected }) => {
-            const value = selected.map(({ value }) => value)
-            onChange({ value }, null)
-        },
-        [onChange]
-    )
+const createChangeHandler = onChange => ({ selected }) => {
+    const value = selected.map(({ value }) => value)
+    onChange({ value }, null)
+}
 
-const useSelectedOptions = (options, value) =>
-    useMemo(() => {
-        const selectedValues = new Set(value || [])
-        return options.filter(({ value }) => selectedValues.has(value))
-    }, [options, value])
+const filterSelectedOptions = (options, value) => {
+    const selectedValues = new Set(value || [])
+    return options.filter(({ value }) => selectedValues.has(value))
+}
 
 const MultiSelectComponent = ({ options, value, onChange, ...rest }) => {
-    const handleChange = useChangeHandler(onChange)
     const renderOptions = Array.isArray(options) ? options : []
-    const selectedOptions = useSelectedOptions(renderOptions, value)
 
     return (
         <MultiSelectField
             {...rest}
-            selected={selectedOptions}
-            onChange={handleChange}
+            selected={filterSelectedOptions(renderOptions, value)}
+            onChange={createChangeHandler(onChange)}
         >
             {renderOptions.map(option => (
                 <MultiSelectOption key={option.value} {...option} />
