+++ src/components/SingleSelect.js
@@ -1,30 +1,21 @@
-import React, { useCallback, useMemo } from 'react'
+import React from 'react'
 import propTypes from '@dhis2/prop-types'
 import { SingleSelectField, SingleSelectOption } from '@dhis2/ui-core'
 
 import { FieldAdapter } from './FieldAdapter.js'
 
-const useChangeHandler = onChange =>
-    useCallback(
-        ({ selected }) => {
-            onChange({ value: selected.value }, null)
-        },
-        [onChange]
-    )
-
-const useSelectedOption = (options, value) =>
-    useMemo(() => options.find(o => o.value === value) || {}, [options, value])
+const createChangeHandler = onChange => ({ selected }) =>
+    onChange({ value: selected.value }, null)
 
 const SingleSelectComponent = ({ options, value, onChange, ...rest }) => {
-    const handleChange = useChangeHandler(onChange)
     const renderOptions = Array.isArray(options) ? options : []
-    const selectedOption = useSelectedOption(renderOptions, value)
+    const selectedOption = renderOptions.find(o => o.value === value) || {}
 
     return (
         <SingleSelectField
             {...rest}
             selected={selectedOption}
-            onChange={handleChange}
+            onChange={createChangeHandler(onChange)}
         >
             {renderOptions.map(option => (
                 <SingleSelectOption key={option.value} {...option} />
