+++ src/components/shared/helpers.js
@@ -1,0 +1,1 @@
+export { createToggleChangeHandler } from './helpers/createToggleChangeHandler.js'
